moved around list components into their own folder | made admin page for adding roles
