--- src/components/forms/AddNomineeForm/index.ts
@@ -1,2 +1,0 @@
-export { default } from './AddNomineeForm';
-export * from './AddNomineeForm';
