--- src/components/forms/AddWinnerForm/index.ts
@@ -1,2 +1,0 @@
-export { default } from './AddWinnerForm';
-export * from './AddWinnerForm';
